requirements.txt
# Core LLM and LangChain dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0  # Multi-proveedor (Claude)
langchain-google-genai>=0.1.0  # Multi-proveedor (Gemini)

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
prompt_toolkit>=3.0.0

# Configuration
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0
sentence-transformers>=2.2.0

# CLI
click>=8.1.0
rich>=13.0.0

# Web API (FastAPI)
fastapi>=0.115.0
uvicorn[standard]>=0.27.0

# SQL parsing and validation
sqlparse>=0.4.4
sqlglot>=23.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
httpx>=0.26.0
sentence-transformers>=2.2.0  # Fase E: ML classification

# Fase F: OpenTelemetry
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0
